requirements.txt
semantic-kernel
python-dotenv
chromadb
pypdf>=3.0.0
pdfplumber
tqdm
python-multipart
